default value added to 'folder' field in attempt to resolve bug #971432; also added traps to prevent PLFNG objects from being created or viewed via ZMI.

+++ PloneLocalFolderNG.py
@@ -20,6 +20,8 @@ from Products.CMFCore.utils import getToolByName
 
 from Acquisition import aq_chain
 
+from App.FindHomes import INSTANCE_HOME   # eg, windows INSTANCE_HOME = C:\Plone\Data
+
 try:
     from Products.mxmCounter import mxmCounter
 except ImportError:
@@ -30,6 +32,7 @@ schema = BaseSchema +  Schema((
     StringField('folder',
                 write_permission=CMFCorePermissions.ManagePortal,
                 required=1,
+                default=INSTANCE_HOME,
                 widget=StringWidget(label='Local directory name')
                 ),
     BooleanField ('require_MD5_with_upload',
@@ -80,6 +83,7 @@ schema = BaseSchema +  Schema((
                 ),
                 
     ))
+
 
 class TypeInfo(SimpleItem):
     """ fake TypeInfo class to make CMF happy """
@@ -284,10 +288,10 @@ class PloneLocalFolderNG(BaseContent):
         if not destfolder.startswith(self.folder):
             raise ValueError('illegal directory: %s' % show_dir)
         
-        zLOG.LOG('PloneLocalFolderNG', zLOG.INFO , "validFolder() :: path = %s" %destfolder )
+        #zLOG.LOG('PloneLocalFolderNG', zLOG.INFO , "validFolder() :: path = %s" %destfolder )
         
         if os.path.exists(destfolder):
-            zLOG.LOG('PloneLocalFolderNG', zLOG.INFO , "validFolder() :: path ok for: %s" %destfolder )
+            #zLOG.LOG('PloneLocalFolderNG', zLOG.INFO , "validFolder() :: path ok for: %s" %destfolder )
             return 1
         else:
             zLOG.LOG('PloneLocalFolderNG', zLOG.INFO , "validFolder() :: !!! path bad for: %s" %destfolder )
@@ -356,34 +360,44 @@ class PloneLocalFolderNG(BaseContent):
             l.append( (d, '%s/%s/plfng_view' % (instance.absolute_url(), '/'.join(sofar))) )
         return l
 
-    def __call__(self, REQUEST, RESPONSE, *args, **kw):
-        rel_dir = '/'.join(REQUEST.get('_e', []))
-        destpath = os.path.join(self.folder, rel_dir)
-
-        if os.path.isfile(destpath):
-            
-            if REQUEST.get('action', '') == 'unpack':
-               self.unpackFile(destpath, REQUEST, RESPONSE)
-            
-            elif REQUEST.get('action', '') == 'delete':
-               self.deleteFile(rel_dir, destpath, REQUEST, RESPONSE)   
-            
-            elif REQUEST.get('action', '') == 'catalog':
-               catalogTool = getToolByName(self, 'portal_catalog')
-               return self.catalogContents()
-            else: 
-               return self.showFile(destpath, REQUEST, RESPONSE)
-        else:
-
-            #  Mozilla browsers don't like backslashes in URLs, so replace any '\' with '/'
-            #  that os.path.join might produce
-            RESPONSE.redirect(('/' + os.path.join(self.absolute_url(1), rel_dir, 'plfng_view')).replace('\\','/'))
+    def __call__(self, REQUEST=None, RESPONSE=None, *args, **kw):
+        if not hasattr(self, "folder"):
+            #zLOG.LOG('PloneLocalFolderNG', zLOG.INFO , "__call__() :: no folder attribute")
+            return self
+        elif not REQUEST:
+            zLOG.LOG('PloneLocalFolderNG', zLOG.INFO , "__call__() :: no REQUEST")
+            raise NotImplementedError, "PLFNG objects can only be created and viewed through the Plone interface."
+        else:     
+           rel_dir = '/'.join(REQUEST.get('_e', []))
+           destpath = os.path.join(self.folder, rel_dir)
+   
+           if os.path.isfile(destpath):
+               
+               if REQUEST.get('action', '') == 'unpack':
+                  self.unpackFile(destpath, REQUEST, RESPONSE)
+               
+               elif REQUEST.get('action', '') == 'delete':
+                  self.deleteFile(rel_dir, destpath, REQUEST, RESPONSE)   
+               
+               elif REQUEST.get('action', '') == 'catalog':
+                  catalogTool = getToolByName(self, 'portal_catalog')
+                  return self.catalogContents()
+               else: 
+                  return self.showFile(destpath, REQUEST, RESPONSE)
+           else:
+   
+               #  Mozilla browsers don't like backslashes in URLs, so replace any '\' with '/'
+               #  that os.path.join might produce
+               RESPONSE.redirect(('/' + os.path.join(self.absolute_url(1), rel_dir, 'plfng_view')).replace('\\','/'))
           
             
     def __bobo_traverse__(self, REQUEST, name, RESPONSE=None):
-
         if not REQUEST.has_key('_e'): 
             REQUEST['_e'] = []
+            
+        if not hasattr(self, "folder"):
+            zLOG.LOG('PloneLocalFolderNG', zLOG.INFO , "__bobo_traverse__() :: no folder attribute")
+            raise NotImplementedError, "This PLFNG object was not properly instantiated.  If it was created through the ZMI, you will have to delete it and add it through the regular Plone interface."
 
         destpath = os.path.join(self.folder, '/'.join(REQUEST['_e']), name)
         if os.path.exists(destpath):
